fastapi_blog_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.46
alembic==1.10.4
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.0
